backend/requirements.txt
flask
flask-cors
boto3
pandas
faiss-cpu
numpy
PyPDF2